Fix ruff errors for dask/array/linalg
